+++ plugin/src/main/java/me/Fupery/ArtMap/Command/CommandHandler.java
@@ -110,7 +110,7 @@ public class CommandHandler implements CommandExecutor {
 									sender.sendMessage(Lang.PREFIX + ChatColor.RED + "No unsaved artwork with that ID.");
 									return;
 								}
-								item = new ArtItem.InProgressArtworkItem(id).toItemStack();
+								item = new ArtItem.InProgressArtworkItem(id, player.getName()).toItemStack();
 							} catch( Exception e) {
 								sender.sendMessage(Lang.PREFIX + ChatColor.RED + "Error retrieving art! Check logs for details.");
 								ArtMap.instance().getLogger().log(Level.SEVERE, "Error retrieving art!", e);